+++ src/main/java/org/taganhorn/FoodCountry/security/UserPrincipal.java
@@ -3,15 +3,15 @@ package org.taganhorn.FoodCountry.security;
 import org.springframework.security.core.AuthenticatedPrincipal;
 
 public class UserPrincipal implements AuthenticatedPrincipal {
-    Integer id;
+    Long id;
     String name;
 
-    public UserPrincipal(Integer id, String name) {
+    public UserPrincipal(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
